Refactor StyleItem component and add authentication to delete style action

+++ src/pages/AddStyle/components/StyleList.jsx
@@ -1,9 +1,12 @@
+import { useLogto } from "@logto/react";
 import axios from "axios";
 import { useEffect, useState } from "react";
+import StyleItem from "./StyleItem";
 import styles from "./styles/StyleList.module.css";
 
 const StyleList = () => {
   const [getAllStyles, setGetAllStyles] = useState([]);
+  const { getAccessToken, isAuthenticated } = useLogto();
 
   useEffect(() => {
     const fetchStyles = async () => {
@@ -21,44 +24,40 @@ const StyleList = () => {
   }, []);
 
   const handleDeleteStyle = async (styleId) => {
-    try {
-      await axios.delete(
-        `${import.meta.env.VITE_APP_API_URL}/style/delete-style/${styleId}`
-      );
-      setGetAllStyles(
-        getAllStyles.filter((style) => style.style_id !== styleId)
-      );
-    } catch (error) {
-      console.error(error.message);
+    if (isAuthenticated) {
+      try {
+        const accessToken = await getAccessToken(
+          `${import.meta.env.VITE_LOGTO_RESOURCES}`
+        );
+        await axios.delete(
+          `${import.meta.env.VITE_APP_API_URL}/style/delete-style/${styleId}`,
+          {
+            headers: {
+              Authorization: `Bearer ${accessToken}`,
+            },
+          }
+        );
+        setGetAllStyles(
+          getAllStyles.filter((style) => style.style_id !== styleId)
+        );
+      } catch (error) {
+        console.error(error.message);
+      }
     }
   };
 
   return (
-    <>
-      <div className={styles["style-list-container"]}>
-        <div className={styles["style-list"]}>
-          {getAllStyles.map((styleItem, index) => (
-            <div key={index} className={styles["style-item"]}>
-              <img
-                src={`${import.meta.env.VITE_APP_ASSET_URL}/${
-                  styleItem.style_img
-                }`}
-                alt={styleItem.style_name}
-                className={styles["style-image"]}
-              />
-              <div className={styles["style-name"]}>{styleItem.style_name}</div>
-              <button
-                className={styles["delete-button"]}
-                onClick={() => handleDeleteStyle(styleItem.style_id)}
-              >
-                X
-              </button>
-            </div>
-          ))}
-        </div>
+    <div className={styles["style-list-container"]}>
+      <div className={styles["style-list"]}>
+        {getAllStyles.map((styleItem, index) => (
+          <StyleItem
+            key={index}
+            styleItem={styleItem}
+            onDelete={handleDeleteStyle}
+          />
+        ))}
       </div>
-    </>
+    </div>
   );
 };
-
 export default StyleList;
